requirements.txt
openai>=1.30.0
playwright>=1.45.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.3
